Extract scene
